requirements_app.txt
# --- Web/API & UI ---
Flask==3.1.1
gunicorn==23.0.0
streamlit==1.36.0

# --- Azure Storage (uploads & blob reads) ---
azure-storage-blob==12.20.0
azure-core==1.30.2
requests==2.32.3

# --- Audio / DSP ---
ffmpeg-python==0.2.0
librosa==0.10.2.post1
soundfile==0.13.1
scipy==1.14.1
numpy==1.26.4
noisereduce==3.0.2

# --- ML / NLP ---
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
transformers==4.52.3
sentencepiece==0.2.0
nltk==3.9.1
langdetect==1.0.9
scikit-learn==1.6.1

# --- YouTube download (Streamlit tab optional) ---
yt-dlp==2024.10.22
moviepy==2.0.0.dev2

# --- Utilities ---
tqdm==4.67.1
regex==2024.11.6
pyyaml==6.0.2
more-itertools==10.7.0
joblib==1.5.0
psutil==7.0.0
platformdirs==4.3.8

# --- Flask internals (usually pulled transitively, pinned for stability) ---
Werkzeug==3.1.3
click==8.2.0
Jinja2==3.1.6
itsdangerous==2.2.0

# --- Whisper (OpenAI) ---
git+https://github.com/openai/whisper.git

# --- Dev (optional) ---
# pytest==8.0.0
# black==24.1.1
# flake8==7.0.0
